+++ Changelog.md
@@ -25,6 +25,10 @@
   (<https://github.com/aws/graph-explorer/pull/419>)
 - Improved testing coverage by having less mocked logic
   (<https://github.com/aws/graph-explorer/pull/421>)
+- Fixed issue where the connection's fetch timeout input would disappear when
+  string is empty (<https://github.com/aws/graph-explorer/pull/445>)
+- Fixed issue where long connection URLs had no vertical padding
+  (<https://github.com/aws/graph-explorer/pull/445>)
 
 ## Release 1.7.0